+++ bindings/ruby/lib/stateset_embedded.rb
@@ -7,5 +7,5 @@ rescue LoadError
 end
 
 module StateSet
-  VERSION = '1.27.0'
+  VERSION = '1.28.0'
 end